Backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
flask-cors
